calico: update to 3.10.2
This is based on templates from
https://github.com/kinvolk/terraform-render-bootkube at commit
0f3156c73b846bf1bc5dd4681879048c37fb2f25.

That project has been used in production a lot, with calico, and took
manifests generated by bootkube initially. Just they were tuned with
calico updates there.

This commit updates all except the daemonset, which is done in a later
commit as the diff is a little more tricky.