calico: update daemonset to 3.10.2
This is based on templates from
https://github.com/kinvolk/terraform-render-bootkube at commit
0f3156c73b846bf1bc5dd4681879048c37fb2f25.

That project has been used in production a lot, with calico, and took
manifests generated by bootkube initially. Just they were tuned with
calico updates there.

This commit updates the daemonset. The tricky part here is that there
are more terraform variables in the terraform-render-bootkube
[file](https://github.com/kinvolk/terraform-render-bootkube/blob/370bac3570505c519b8db1113664439d4c57a0e8/resources/calico/daemonset.yaml).

This way is easier to review, but basically all terraform variables were
replaced by they default value in the correspondig variables
[file](https://github.com/kinvolk/terraform-render-bootkube/blob/370bac3570505c519b8db1113664439d4c57a0e8/variables.tf)