assets: Add comment to avoid mistakes
CalicoNodeTemplate asset is used for calico while
CalicoPolicyOnlyTemplate is used for canal (IIUC, to get the policy only
from calico and use flannel for other stuff).

However, the content of both look very similar, names are not obvious and
is easy to confuse them during changes. Hopefully, this comment helps
until canal is dropped.